+++ CHANGELOG.md
@@ -16,6 +16,14 @@
 
 - <PR-#ISSUE> ...
 
+## `1.0.0` - 7/5/2023
+
+#### Added
+
+- <PR-#20> Add Magic Connect Admin SDK support for Token Resource
+  - [Security Enhancement]: Validate `aud` using Magic client ID. 
+  - Pull client ID from Magic servers if not provided in constructor.
+
 ## `0.2.0` - 11/23/2022
 
 #### Added